LPS-120269 Move mapper to petra - move test
