+++ modules/core/petra/petra-url-pattern-mapper/src/test/java/com/liferay/petra/url/pattern/mapper/internal/DynamicSizeTrieURLPatternMapperPerformanceTest.java
@@ -12,35 +12,17 @@
  * details.
  */
 
-package com.liferay.portal.remote.cors.internal.url.pattern.mapper;
+package com.liferay.petra.url.pattern.mapper.internal;
 
-import java.util.HashMap;
+import com.liferay.petra.url.pattern.mapper.URLPatternMapper;
+
 import java.util.Map;
-
-import org.junit.Assert;
-import org.junit.Test;
 
 /**
  * @author Arthur Chan
  */
-public class DynamicSizeTrieURLPatternMapperTest
-	extends SimpleURLPatternMapperTest {
-
-	@Test
-	public void testConstructor() {
-		Map<String, String> map = new HashMap<>();
-
-		for (int i = 0; i < 1024; i++) {
-			map.put("*.key" + i, "value" + i);
-		}
-
-		URLPatternMapper<String> urlPatternMapper = createURLPatternMapper(map);
-
-		for (int i = 0; i < 1024; i++) {
-			Assert.assertEquals(
-				"value" + i, urlPatternMapper.getValue("*.key" + i));
-		}
-	}
+public class DynamicSizeTrieURLPatternMapperPerformanceTest
+	extends SimpleURLPatternMapperPerformanceTest {
 
 	@Override
 	protected URLPatternMapper<String> createURLPatternMapper(
