+++ modules/core/petra/petra-url-pattern-mapper/src/test/java/com/liferay/petra/url/pattern/mapper/internal/SimpleURLPatternMapperTest.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.remote.cors.internal.url.pattern.mapper;
+package com.liferay.petra.url.pattern.mapper.internal;
 
+import com.liferay.petra.url.pattern.mapper.URLPatternMapper;
 import com.liferay.portal.kernel.test.rule.CodeCoverageAssertor;
 
 import java.util.HashMap;
